% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decp_plot.R
\name{decp_plot}
\alias{decp_plot}
\title{Descriptive Plot}
\usage{
decp_plot(x)
}
\arguments{
\item{x}{the data-set you want to plot}
}
\value{
a 2 x 2 gragh with 4 plots showing descriptive analysis
}
\description{
Plot the descriptive plots for certain dataset including histogram,
density plot, Quantile Plot and Boxplot.
}
\examples{
data <- read.csv("data.csv")
decp_plot(data)

}
