% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_series.R
\name{fred_get_series}
\alias{fred_get_series}
\title{Get Series Data from FRED}
\usage{
fred_get_series(APIkey, series_id, observation_start)
}
\arguments{
\item{APIkey}{Your API key for FRED( which could get freely from FRED)}

\item{series_id}{the specific data you want to require from FRED}

\item{observation_start}{the start time of the observation}
}
\value{
a data frame with date and required value from the observation date you give
}
\description{
Get Series Data from FRED
}
\examples{
fred_get_series("API_KEY",c("GNPCA", "CPIAUCSL"),"1990-01-01")

}
