% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_APOD.R
\name{gene_APOD}
\alias{gene_APOD}
\title{Get Astronomy Picture of the Day from NASA}
\usage{
gene_APOD(date)
}
\arguments{
\item{date}{which should be in date format}
}
\value{
a data frame incluing all the information including explanation, picture url, author and so on.
}
\description{
Get the information about the astronomy picture of the day you select from NASA.
}
\examples{
gene_APOD("2019-01-01")

}
