% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_plot.R
\name{irf_plot}
\alias{irf_plot}
\title{irf Plot}
\usage{
irf_plot(y_model, k)
}
\arguments{
\item{y_model}{This is a VAR model for  prediction.}

\item{k}{predict k steps ahead, for example, 12 or 36.}
}
\value{
4 plot
}
\description{
To plot prediction Impulse response function in 4 separate graphs
in case the two variables have toally different y-scale.
}
\examples{
irf_plot(model,36)

}
