% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAData.R
\docType{data}
\name{NBA_Center}
\alias{NBA_Center}
\title{NBA Stat}
\format{
110 X 30 data frame each
}
\usage{
NBA_Center
}
\description{
the stat for all NBA player in 2019, ordering by their age and classifying by their position.
}
\examples{
"NBA_Center"


}
\keyword{datasets}
